% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumprodRcpp}
\alias{cumprodRcpp}
\title{cumprodRcpp}
\usage{
cumprodRcpp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
c++ version of base cumprod(). Missing values not accepted.
}

