% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diffRcpp}
\alias{diffRcpp}
\title{diffRcpp}
\usage{
diffRcpp(x, lag = 1L)
}
\arguments{
\item{x}{a numeric vector}

\item{lag}{the specified lag}
}
\description{
c++ version of base diff(). Missing values not accepted.
}

