% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxRcpp}
\alias{maxRcpp}
\title{maxRcpp}
\usage{
maxRcpp(x, na_rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na_rm}{boolean, should NA's be considered?}
}
\description{
c++ version of base max(). can deal with missing values
}

