% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxRcpp2}
\alias{maxRcpp2}
\title{maxRcpp2}
\usage{
maxRcpp2(x, na_rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na_rm}{boolean, should NA's be considered?}
}
\description{
c++ version of base max(). Different implementation.
  can deal with missing values. Its not as fast as maxRcpp()
}

