% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rangeRcpp}
\alias{rangeRcpp}
\title{rangeRcpp}
\usage{
rangeRcpp(x, na_rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na_rm}{boolean, should NA's be considered?}
}
\description{
c++ version of base range(), can deal with na's
}

